Fix migration script
