+++ src/zenml/models/v2/misc/server_models.py
@@ -17,7 +17,7 @@ from uuid import UUID, uuid4
 
 from pydantic import BaseModel, Field
 
-from zenml.enums import AuthScheme, SecretsStoreType
+from zenml.enums import AuthScheme
 from zenml.utils.enum_utils import StrEnum
 
 
@@ -65,10 +65,6 @@ class ServerModel(BaseModel):
     database_type: ServerDatabaseType = Field(
         ServerDatabaseType.OTHER,
         title="The database type that the server is using.",
-    )
-    secrets_store_type: SecretsStoreType = Field(
-        SecretsStoreType.NONE,
-        title="The type of secrets store that the server is using.",
     )
     auth_scheme: AuthScheme = Field(
         title="The authentication scheme that the server is using.",
